+++ datasource/netcdf.py
@@ -41,7 +41,8 @@ DISCHARGE_LATERAL = NcVar('qlat', 'discharge lateral', 'm3/s')
 VELOCITY_INTERFLOW = NcVar('up1', 'velocity interflow', 'm/s')
 RAIN_INTENSITY = NcVar('rain', 'rain intensity', 'm3/s')
 WET_SURFACE_AREA = NcVar('su', 'wet surface area', 'm2')
-INFILTRATION = NcVar('infiltration', 'infiltration rate', 'm3/s')
+INFILTRATION = NcVar('infiltration_rate', 'infiltration rate', 'm3/s')
+WET_CROSS_SECTION_AREA = NcVar('au', 'wet surface area', 'm2')
 
 _Q_TYPES = [
     DISCHARGE,
@@ -49,6 +50,7 @@ _Q_TYPES = [
     DISCHARGE_PUMP,
     VELOCITY,
     VELOCITY_INTERFLOW,
+    WET_CROSS_SECTION_AREA,
 ]
 
 _H_TYPES = [